Refactors the Entity struct into a WebResource interface.
The Entity struct is clumsy for our purposes:
- Structs do not inherit from each other in an object-oriented sense.  A struct that embeds the Entity struct is *not* of type `Entity`.  We wish to perform type assertions, and embedded structs are not a path forward for that.  Interfaces are.
- "Entity" is a bad name, it implies (to my mind) some sort of relationship with the persistence layer, and this is not true.  The persistence layer will store and retrieve anything that uses "id" as the key for storing the object's key.  What we really mean is that this is a web-based resource, and these fields on "Entity" are being used in the HTTP layer to address web concerns like caching and emit proper HTTP headers

Therefore, the use of embedded structs are removed and replaced with an interface named WebResource.

Model structs that require treatment as web resources will need to implement the WebResource interface, which at this juncture has some ugliness with it, because the common fields are not captured in an embedded struct.